requirements.txt
langchain_community
deep_translator
flask_cors
langchain_openai
streamlit
chromadb
langchain-community
pypdf
unstructured
openai
transformers
langchain
google-generativeai
protobuf
python-dotenv
